+++ tests/at_functional_test/test/enroll_verb_test.dart
@@ -254,7 +254,7 @@ void main() {
       Map llookupResponseMap = jsonDecode(llookupResponse);
       expect(llookupResponseMap['errorCode'], 'AT0009');
       expect(llookupResponseMap['errorDescription'],
-          'UnAuthorized client in request : Enrollment Id: $enrollmentId is not authorized for local lookup operation on the key: $enrollmentKey');
+          'UnAuthorized client in request : Connection with enrollment ID $enrollmentId is not authorized to llookup key: $enrollmentKey');
 
       // keys:get:self should return default self encryption key
       var selfKey = '$enrollmentId.default_self_enc_key.__manage$firstAtSign';
@@ -284,6 +284,48 @@ void main() {
           privateKeyGetResponse.contains(
               '${apkamEncryptedKeysMap['encryptedDefaultEncPrivateKey']}'),
           true);
+    });
+
+    test(
+        'enroll request on APKAM authenticated connection and verify enroll:list',
+        () async {
+      int randomNumber = Uuid().v4().hashCode;
+      await firstAtSignConnection.authenticateConnection(
+          authType: AuthType.pkam);
+      // send an enroll request with the keys from the setEncryptionKeys method
+      String enrollRequest =
+          'enroll:request:{"appName":"atmosphere-$randomNumber","deviceName":"pixel-$randomNumber","namespaces":{"wavi":"rw"},"encryptedDefaultEncryptionPrivateKey":"${apkamEncryptedKeysMap['encryptedDefaultEncPrivateKey']}","encryptedDefaultSelfEncryptionKey":"${apkamEncryptedKeysMap['encryptedSelfEncKey']}","apkamPublicKey":"${pkamPublicKeyMap[firstAtSign]!}","encryptedAPKAMSymmetricKey":"dummy_apkam_$randomNumber"}';
+      String enrollResponse =
+          (await firstAtSignConnection.sendRequestToServer(enrollRequest))
+              .replaceFirst('data:', '');
+      var enrollJsonMap = jsonDecode(enrollResponse);
+      var enrollmentId = enrollJsonMap['enrollmentId'];
+      expect(enrollmentId, isNotEmpty);
+      expect(enrollJsonMap['status'], 'pending');
+      // enroll:list
+      String enrollListResponse =
+          await firstAtSignConnection.sendRequestToServer('enroll:list');
+      enrollListResponse = enrollListResponse.replaceAll('data:', '');
+      var enrollListResponseMap = jsonDecode(enrollListResponse);
+      expect(
+          enrollListResponseMap[
+              '$enrollmentId.new.enrollments.__manage$firstAtSign']['appName'],
+          'atmosphere-$randomNumber');
+      expect(
+          enrollListResponseMap[
+                  '$enrollmentId.new.enrollments.__manage$firstAtSign']
+              ['deviceName'],
+          'pixel-$randomNumber');
+      expect(
+          enrollListResponseMap[
+          '$enrollmentId.new.enrollments.__manage$firstAtSign']
+          ['namespace']['wavi'],
+          'rw');
+      expect(
+          enrollListResponseMap[
+                  '$enrollmentId.new.enrollments.__manage$firstAtSign']
+              ['encryptedAPKAMSymmetricKey'],
+          'dummy_apkam_$randomNumber');
     });
 
     // Purpose of the tests
@@ -403,11 +445,13 @@ void main() {
               serverResponse.contains(
                   '${enrollJson['enrollmentId']}.new.enrollments.__manage'),
               true);
-          Map notificationValue = jsonDecode(jsonDecode(serverResponse.replaceAll('notification: ',''))['value']);
+          Map notificationValue = jsonDecode(jsonDecode(
+              serverResponse.replaceAll('notification: ', ''))['value']);
           expect(notificationValue['appName'], 'buzz');
           expect(notificationValue['deviceName'], 'pixel');
-          expect(notificationValue['namespace'], {'buzz':'rw'});
-          expect(notificationValue['encryptedApkamSymmetricKey'], apkamEncryptedKeysMap['encryptedApkamSymmetricKey']);
+          expect(notificationValue['namespace'], {'buzz': 'rw'});
+          expect(notificationValue['encryptedApkamSymmetricKey'],
+              apkamEncryptedKeysMap['encryptedApkamSymmetricKey']);
           monitorSocket.close();
         }
         /* Setting count to 4 to wait until server returns 4 responses